requirements.txt
altair==5.5.0
asgiref==3.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.60.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
narwhals==2.5.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
paystackapi==2.1.3
pillow==11.3.0
protobuf==6.32.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
